Use the :generate-virtual-row event for simulating new rows. (#130)
Previously the :parse-query event did not call the :generate-virtual-row
event. The new row simulation happened inside the parsing code.

Now, the parsing code emits a map that triggers a call to
:generate-virtual-row.

Newly generated rows and simulations for cells are removed whenever
there are negative values. This hack was used when generating rows
inside the parse query code. Now it is just being brought into the code
in :generate-virtual-row and the sub :generator.