Add ability to trigger application of custom cell renderer functions (#135)
* Add new file with cell renderer functions
* Create subscriptions that select a renderer function based on which
confidence mode is selected.

Currently the same rendering effect is applied no matter which
confidence mode is selected because we do not have real data in the
likelihoods subscriptions. However, this is putting all the scaffolding
in place for when we do have valid data coming out of those
subscriptions.