Update the format of data returned by cell-wise scoring functions.
The changes are intended to make the data returned more sensible.

Also updated db specs because values returned by the
cell-wise scoring functions get stored in the db.

Also in the code edited, references to likelihood were changed to
score, as that term is more context free.

This change breaks the ability to show imputed values and color cells by
confidence using the confidence mode menu.