Update schema versions for vega-lite specs
Some specs have been updated to the v4-beta schema. And some specs have
been been kept at v3. The ones upgraded to v4 are the specs currently in
use by the app and have been confirmed to work. Upgrading the specs is
necessary to supress warning messages about running specs with old
schemas in a newer version of vega-lite.

The specs that have been set to v3 are one that are not currently in use
by the app and have not been tested with the latest version of vega-lite.
This is to retain the warning message when runnning these specs in case
there is any weird behaviour with the new version of vega-lite.