Make the log-level for the vega-embed component dynamic.
The log-level is now a subscription. The log-level is decreased when a
single-cell is selected. In this case, only errors get printed to the
console and not warnings.

This was done to supress warning messages complaining about infinite
scale extents. This is likely due to the plot starting out with no data
combined with the addition of more warning messages in the newest version of
vega that has been recently added.

This change does not change the warning level when more than a single
cell is selected.