Use a vector of objects for row data in handsontable.
Previously we were using a vector of vectors for rows in the table.

This change involved creating a helper function that produces settings for
all the columns. Mostly, specifying which data attribute to select for
each column.

This change may make detecting user interactions in columns easier in the
future because we will be able to deal with data attributes instead of
indicies within callback functions.