Allow selection of non-adjacent columns for visualization.
Previously in the spreadsheets app, the user had to move two columns
right next to each other in order select both of them for visualization.

Now, the user can select a pair of columns that are non-adjacent
by holding down the control(command) key before selecting the second
one.

Changes that have been made.

* Multiple selection mode has been enabled for Hansontable.

* The :after-selection-end event has been refactored extensively. The
event is now an fx event instead of a db event, and it sometimes calls a
new re-frame effect that has been added which deselects all cells. The
other major thing that has changed is that we only update the db when
the selection is valid.

* We now emit a nil when we do not want to display any vega-lite spec.
Emitting an empty map caused vega-lite to emit an error. This case for
not displaying any spec was hit often during some of new situations
introduced by the changes in this commit.