Refactor :control/set-reagent-forms event.
The event is now a case statement with special cases for writing into
certain paths.

The :control/set-selection-color case prevents writing nil values for
:selection-color.

The :control/set-confidence-mode case is mostly the same as before when
writing to :confidence mode. However, there is the addition of some
accessor functions from the highlight component's db. Before the keys
we were checking to determine if we needed to run a computation event were
outdated (always nil). This fixes that by using accessor functions that
check the correct, updated path.