Add query plan interpreter (#204)
Allows us to enhance Datalog with our own custom forms by implementing an interpreter for a customizable superset of Datalog. This interpreter alternates between delegating to Datascript and running its own custom operations, using a relation of all the variables in the query to keep track of the "query state".