Remove :hot/after-deselect event.
Upcoming changes will make it so that it will not be used to know when
to clear out selections in the db. Instead it will work the other way
around. Setting the selection state to nil in the db will clear
selections in the table causing a :after-deselect event to be emitted.
An event which we will no longer be hooked into.