Allow pan/zoom for scatter plots and strip plots. (#233)
Removed autosize-resize setting from final vega-lite spec. This resize
setting is needed for simulation plots to resize dynamically. However
this setting causes jitter when other plots are panned or zoomed.

The stop-gap solution here is to resize the plot whenever new data is
streamed in for the simulation plots. This does still cause some jitter
when there is a simulation plot present and an other chart is panned and
zoomed. However, this is a working solution for now until simulation
plots can be rendered in their own vega-embed instance.