Enhance publish script and makefile target for publish.
Adds a new makefile target, publish-dir, which sets up the publish
directory. This work was previously being done by the publish shell
script. This target also now copies a CNAME file and AUTH file if they
exist.

The target for publish can now take an argument for a domain that it
will pass to the publish shell script.

The publish script itself is now more robust. It can now safely be
called from any directory although the script itself must remain in the
bin/ directory.

The publish script also requires that an AUTH file is present before publishing.