Fix selections in filtered tables. (#249)
Previously when the table was filtered, the new table state was not
updated in the re-frame db. This commit adds a new handsontable event
hook that does this updating.

Now when selections are made on filtered data, the
:table/check-selection event compares the selection rectangles against
the correct number of rows in the table, and does not incorrectly cancel
the selection.