Remove config.edn from .gitignore (#287)
This is to help the user see that they have a config.edn file in the
project root that is getting read by the app. Often times a config.edn
file in the root dir is not noticed, and it is difficult to figure out
why the app is not compiling.