Save changes to table into app-db.
The :hot/after-change event now saves changes to editable cells.

The changes are first validated based on the datatype in that column,
and thy are then cast accordingly (from string to number if needed).

The changes are then merged into the datasets—both the original dataset
and the currently displayed dataset.