Change setting for Figwheel css dir (#364)
This new directory points at the css folder that is created for
Figwheel instead of the one in the source tree. This will make live
css-reloading more sensible for Figwheel users.

Figwheel will hot-reload css when the css file used by the Figwheel app
instance is changed not the one in the source tree. This may require
some manually copying or use of a script by the Figwheel user.