Reorganize css and icons in resources directory (#367)
Created new sub-directories within the resources directory for css,
libs, and icons.

Moved css and icons into respective new dirs.

Added the full Bootstrap css file as a dependency instead of the just
the partial one we were depending on before. This full css file will be
needed by the Re-com library which will be added in a later commit.