Logging various subs that might be useful
Cameron: you will likely find this shelpful.

Do use this model object which is an XCat record, instead of the bayes-db-dump you were using before. It should have all you need. You can treat it like a clojure map or use some of the other functions in iql.inference to manipulate it.

Also be sure to have clojurescript formatters installed in chrome if not done already.